refactor: Place Controller Swagger 변경 및 PlaceService에 멤버권한 체크 기능 추가

+++ backend/src/main/java/middle_point_search/backend/domains/place/service/PlaceService.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import middle_point_search.backend.common.exception.CustomException;
 import middle_point_search.backend.domains.member.domain.Member;
-import middle_point_search.backend.domains.member.service.MemberService;
+import middle_point_search.backend.domains.memberRoom.MemberRoomValidateService;
 import middle_point_search.backend.domains.place.domain.Place;
 import middle_point_search.backend.domains.place.dto.PlaceDTO.PlaceSaveOrUpdateRequest;
 import middle_point_search.backend.domains.place.dto.PlaceDTO.PlaceVO;
@@ -28,12 +28,15 @@ import middle_point_search.backend.domains.room.service.RoomService;
 public class PlaceService {
 
 	private final PlaceRepository placeRepository;
-	private final MemberService memberService;
 	private final RoomService roomService;
+	private final MemberRoomValidateService memberRoomValidateService;
 
 	//장소 저장
 	@Transactional(rollbackFor = {CustomException.class})
 	public void savePlace(Long roomId, Member member, PlaceSaveOrUpdateRequest request) {
+		// 회원이 방에 속해있는지 확인
+		memberRoomValidateService.validateAuthorizedMember(member.getId(), roomId);
+
 		Room room = roomService.findRoom(roomId)
 			.orElseThrow(() -> CustomException.from(ROOM_NOT_FOUND));
 
@@ -41,7 +44,10 @@ public class PlaceService {
 	}
 
 	// 장소 조회
-	public PlacesFindResponse findPlaces(Long roomId) {
+	public PlacesFindResponse findPlaces(Long memberId, Long roomId) {
+		// 회원이 방에 속해있는지 확인
+		memberRoomValidateService.validateAuthorizedMember(memberId, roomId);
+
 		List<Place> places = placeRepository.findAllByRoom_Id(roomId);
 
 		List<PlaceVO> placeVOs = places.stream()
@@ -54,7 +60,10 @@ public class PlaceService {
 
 	// 장소 삭제
 	@Transactional
-	public void deletePlace(Long placeId) {
+	public void deletePlace(Long memberId, Long placeId) {
+		// 회원이 방에 속해있는지 확인
+		memberRoomValidateService.validateAuthorizedMember(memberId, placeId);
+
 		placeRepository.deleteById(placeId);
 	}
 }
